app/core/config.py
import os
from dotenv import load_dotenv

# 환경 변수 로드
load_dotenv()

class Settings:
    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY")
    PINECONE_API_KEY: str = os.getenv("PINECONE_API_KEY")

settings = Settings()