pyproject.toml
[tool.poetry]
name = "ml-backend-python"
version = "0.1.0"
description = ""
authors = ["an-jiohh <dkswlgh99@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
pinecone-client = "^5.0.1"
torch = "^2.4.1"
kobert-transformers = "^0.6.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
